+++ packages/domain/src/storyboard/index.ts
@@ -2,3 +2,5 @@ export * from "./types";
 export * from "./ports";
 export * from "./eligibility";
 export * from "./fingerprint";
+export * from "./ordering";
+export * from "./duration";